Added missing test for RateLimitServiceTest && refactoring
